.github/scripts/requirements.txt
google-genai
openai
requests
qdrant-client